evals/requirements.txt
mlflow>=2.20
databricks-agents>=0.10
databricks-sdk>=0.105,<1
